requirements.txt
yfinance
fredapi
requests
pandas
mysql-connector-python
psycopg2-binary
pymongo
streamlit
matplotlib
anthropic
statsmodels
